# api/openapi.yaml
openapi: 3.0.3
info:
  title: Kafkas Derneği Panel API
  description: Kafkas Derneği yönetim paneli API dokümantasyonu
  version: 1.0.0
  contact:
    name: API Support
    email: support@kafkasder.org
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.kafkasder.org/api
    description: Production server
...
paths:
...
  /social-aid:
    get:
      tags:
        - Social Aid
      summary: Sosyal yardım başvuru listesi
      description: Tüm sosyal yardım başvurularını sayfalı olarak getirir
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: durum
          in: query
          schema:
            type: string
            description: Duruma göre filtrele
      responses:
        '200':
          description: Başarılı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        '500':
          description: Sunucu hatası
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Social Aid
      summary: Yeni sosyal yardım başvurusu oluştur
      description: Sisteme yeni sosyal yardım başvurusu ekler
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SocialAidApplicationInput'
      responses:
        '201':
          description: Başvuru başarıyla oluşturuldu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SocialAidApplication'
        '400':
          description: Geçersiz veri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Hata mesajı
        message:
          type: string
          example: Detaylı hata açıklaması
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items: {}
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
...
    SocialAidApplicationInput:
      type: object
      required:
        - beneficiary_id
        - yardim_turu
        - gerekce
      properties:
        beneficiary_id:
          type: number
        yardim_turu:
          type: string
          enum: [gida, egitim, saglik, barinma, giyim, diger]
        talep_edilen_tutar:
          type: number
        gerekce:
          type: string
          minLength: 10
        durum:
          type: string
          enum: [beklemede, inceleniyor, onaylandi, reddeddi]
...
    SocialAidApplication:
      allOf:
        - $ref: '#/components/schemas/SocialAidApplicationInput'
      type: object
      properties:
        id:
          type: number
        basvuru_tarihi:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...